exception handling - delete

+++ src/main/java/com/projetoSpring/course/services/UserService.java
@@ -4,11 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.projetoSpring.course.entities.User;
 import com.projetoSpring.course.repositories.UserRepository;
+import com.projetoSpring.course.services.exceptions.DataBaseException;
 import com.projetoSpring.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -32,8 +34,17 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
-	}
+	  if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException(id);
+	  }
+		 try{
+	       repository.deleteById(id);
+		 }
+	     catch(DataIntegrityViolationException e) {
+		   throw new DataBaseException(e.getMessage());
+	     }
+	  
+	 }
 	
 	@Transactional
 	public User update(Long id, User obj) {
